requirements.txt
# PPTX Builder - Runtime Dependencies
# Install with: pip install -r requirements.txt

python-pptx>=0.6.21
Pillow>=10.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pillow-heif>=0.13.0
